+++ funnel/forms/account.py
@@ -1,7 +1,3 @@
-from collections import namedtuple
-
-from flask import request
-
 from baseframe import _, __
 from coaster.auth import current_auth
 from coaster.utils import nullstr, sorted_timezones
@@ -15,10 +11,8 @@ from ..models import (
     UserPhone,
     UserPhoneClaim,
     getuser,
-    notification_type_registry,
     password_policy,
 )
-from ..transports import platform_transports
 from ..utils import strip_phone, valid_phone
 from .helpers import EmailAddressAvailable
 
@@ -30,8 +24,6 @@ __all__ = [
     'PasswordResetForm',
     'PasswordChangeForm',
     'AccountForm',
-    'UnsubscribeForm',
-    'SetNotificationPreferenceForm',
     'EmailPrimaryForm',
     'ModeratorReportForm',
     'NewEmailAddressForm',
@@ -41,7 +33,6 @@ __all__ = [
     'VerifyPhoneForm',
     'supported_locales',
     'timezone_identifiers',
-    'transport_labels',
 ]
 
 
@@ -51,72 +42,6 @@ timezone_identifiers = dict(timezones)
 supported_locales = {
     'en': __("English"),
     'hi': __("Hindi (beta; incomplete)"),
-}
-
-TransportLabels = namedtuple(
-    'TransportLabels',
-    [
-        'title',
-        'requirement',
-        'unsubscribe_form',
-        'switch',
-        'enabled_main',
-        'enabled',
-        'disabled_main',
-        'disabled',
-    ],
-)
-transport_labels = {
-    'email': TransportLabels(
-        __("Email"),
-        __("To enable, add a verified email address"),
-        __("Notify me by email"),
-        __("Email notifications"),
-        __("Enabled selected email notifications"),
-        __("Enabled this email notification"),
-        __("Disabled all email notifications"),
-        __("Disabled this email notification"),
-    ),
-    'sms': TransportLabels(
-        __("SMS"),
-        __("To enable, add a verified phone number"),
-        __("Notify me by SMS"),
-        __("SMS notifications"),
-        __("Enabled selected SMS notifications"),
-        __("Enabled this SMS notification"),
-        __("Disabled all SMS notifications"),
-        __("Disabled this SMS notification"),
-    ),
-    'webpush': TransportLabels(
-        __("Browser"),
-        __("To enable, allow push notifications in the browser"),
-        __("Notify me with browser notifications"),
-        __("Push notifications"),
-        __("Enabled selected push notifications"),
-        __("Enabled this push notification"),
-        __("Disabled all push notifications"),
-        __("Disabled this push notification"),
-    ),
-    'telegram': TransportLabels(
-        __("Telegram"),
-        __("To enable, link your Telegram account"),
-        __("Notify me on Telegram"),
-        __("Telegram notifications"),
-        __("Enabled selected Telegram notifications"),
-        __("Enabled this Telegram notification"),
-        __("Disabled all Telegram notifications"),
-        __("Disabled this Telegram notification"),
-    ),
-    'whatsapp': TransportLabels(
-        __("WhatsApp"),
-        __("To enable, add your WhatsApp number"),
-        __("Notify me on WhatsApp"),
-        __("WhatsApp notifications"),
-        __("Enabled selected WhatsApp notifications"),
-        __("Enabled this WhatsApp notification"),
-        __("Disabled all WhatsApp notifications"),
-        __("Disabled this WhatsApp notification"),
-    ),
 }
 
 
@@ -393,121 +318,6 @@ class AccountForm(forms.Form):
             raise forms.ValidationError(_("This username has been taken"))
         else:
             raise forms.ValidationError(_("This username is not available"))
-
-
-@User.forms('unsubscribe')
-class UnsubscribeForm(forms.Form):
-    __expects__ = ('edit_user', 'transport', 'notification_type')
-
-    # To consider: Replace the field's ListWidget with a GroupedListWidget, and show all
-    # known notifications by category, not just the ones the user has received a
-    # notification for. This will avoid a dark pattern wherein a user keeps getting
-    # subscribed to new types of notifications, a problem Twitter had when they
-    # attempted to engage dormant accounts by inventing new reasons to email them.
-    # However, also consider that this will be a long and overwhelming list, and will
-    # not help with new notification types added after the user visits this list. The
-    # better option may be to set notification preferences based on previous
-    # preferences. A crude form of this exists in the NotificationPreferences class,
-    # but it should be smarter about defaults per category of notification.
-
-    main = forms.BooleanField(
-        __("Notify me"), description=__("Uncheck this to disable all notifications"),
-    )
-
-    types = forms.SelectMultipleField(
-        __("Or disable only a specific notification"),
-        widget=forms.ListWidget(),
-        option_widget=forms.CheckboxInput(),
-    )
-
-    # This token is validated in the view, not here, because it has to be valid in the
-    # GET request itself, and the UI flow is very dependent on the validation error.
-    token = forms.HiddenField(
-        __("Unsubscribe token"), validators=[forms.validators.DataRequired()]
-    )
-    token_type = forms.HiddenField(
-        __("Unsubscribe token type"), validators=[forms.validators.DataRequired()]
-    )
-
-    def set_queries(self):
-        # Populate choices with all notification types that the user has a preference
-        # row for.
-        if self.transport in transport_labels:
-            self.main.label.text = transport_labels[self.transport].unsubscribe_form
-        self.types.choices = [
-            (ntype, notification_type_registry[ntype].description)
-            for ntype in self.edit_user.notification_preferences
-            if ntype in notification_type_registry
-            and notification_type_registry[ntype].allow_transport(self.transport)
-        ]
-        # Sort by type string. Temp fix for stable ordering until we introduce grouping
-        self.types.choices.sort()
-
-        if request.method == 'GET':
-            # Populate data with all notification types for which the user has the
-            # current transport enabled
-            self.main.data = self.edit_user.main_notification_preferences.by_transport(
-                self.transport
-            )
-            self.types.data = [
-                ntype
-                for ntype, user_prefs in self.edit_user.notification_preferences.items()
-                if user_prefs.by_transport(self.transport)
-            ]
-
-    def save_to_user(self):
-        # self.types.data will only contain the enabled preferences. Therefore, iterate
-        # through all choices and toggle true or false based on whether it's in the
-        # enabled list. This uses dict access instead of .get because set_queries
-        # also populates from this list.
-        self.edit_user.main_notification_preferences.set_transport(
-            self.transport, self.main.data
-        )
-        for ntype, title in self.types.choices:
-            self.edit_user.notification_preferences[ntype].set_transport(
-                self.transport, ntype in self.types.data
-            )
-
-
-@User.forms('set_notification_preference')
-class SetNotificationPreferenceForm(forms.Form):
-    """Set one notification preference."""
-
-    notification_type = forms.SelectField(__("Notification type"))
-    transport = forms.SelectField(
-        __("Transport"), validators=[forms.validators.DataRequired()],
-    )
-    enabled = forms.BooleanField(__("Enable this transport"))
-
-    def set_queries(self):
-        # The main switch is special-cased with an empty string for notification type
-        self.notification_type.choices = [('', __("Main switch"))] + [
-            (ntype, cls.description)
-            for ntype, cls in notification_type_registry.items()
-        ]
-        self.transport.choices = [
-            (transport, transport)
-            for transport in platform_transports
-            if platform_transports[transport]
-        ]
-
-    def status_message(self):
-        """Render a success or error message."""
-        if self.errors:
-            # Flatten errors into a single string because typically this will only
-            # be a CSRF error.
-            return ' '.join(' '.join(message) for message in self.errors.values())
-        if self.notification_type.data == '':
-            return (
-                transport_labels[self.transport.data].enabled_main
-                if self.enabled.data
-                else transport_labels[self.transport.data].disabled_main
-            )
-        return (
-            transport_labels[self.transport.data].enabled
-            if self.enabled.data
-            else transport_labels[self.transport.data].disabled
-        )
 
 
 def validate_emailclaim(form, field):
